Refactored exercise results and their totals
