+++ src/components/Exercise/exercise-results.jsx
@@ -3,6 +3,7 @@ import {clearAuth} from '../../actions/auth';
 import {clearAuthToken} from '../../local-storage';
 import NutritionSearchPage from '../nutrition-search-page';
 import ExerciseSearchPage from './exercise-search-page';
+import ExerciseResultsTotals from './exercise-results-totals';
 import {connect} from 'react-redux';
 import {addExercise} from '../../actions/protected-exercise-data';
 import requiresLogin from '../Login/requires-login';
@@ -46,36 +47,6 @@ export class ExerciseResults extends Component {
     e.preventDefault();
     this.props.dispatch(addExercise(cal, dur, username, date));
     window.alert('You have just saved your exercise!');
-  }
-
-  renderTotals() {
-    let exercise_results_array = this.props.exerciseResults;
-    let exerciseTotals = {
-      'name' : '',
-      'nf_calories' : 0,
-      'duration_min' : 0
-    };
-    let i_count;
-    
-    for (let i = 0; i < exercise_results_array.length; i++) {
-      let name = exercise_results_array[i].name.toLowerCase().split(' ').map(s => s.charAt(0).toUpperCase() + s.substring(1)).join(' ');
-      exerciseTotals.name = exerciseTotals.name += name + (exercise_results_array[i + 1] ? ', ' : '');
-      exerciseTotals.nf_calories = Math.floor(exerciseTotals.nf_calories += exercise_results_array[i].nf_calories);
-      exerciseTotals.duration_min = Math.floor(exerciseTotals.duration_min += exercise_results_array[i].duration_min);
-      i === exercise_results_array.length - 1 ? i_count = i : console.log('i error');
-    }
-  
-    if (i_count >= 0) {
-      return (
-        <div className = {!this.props.loading ? "exercise-totals fadeIn" : "exercise-totals"}>
-          <form onSubmit = {(e) => this.onAdd(e, exerciseTotals.name, exerciseTotals.nf_calories, exerciseTotals.duration_min, this.props.username, Date.now)}>
-            <h4>Total Calories Burned : {exerciseTotals.nf_calories}</h4>
-            <h4>Total Duration : {exerciseTotals.duration_min} minutes</h4>
-            <button type = "submit" className="save-btn">Save Exercise</button>
-          </form>
-        </div>
-      );
-    }
   }
 
   render() {
@@ -123,7 +94,7 @@ export class ExerciseResults extends Component {
         <ul className={!this.props.loading ? "exercise-results fadeIn" : "exercise-results"}>
           {exercise_result}
         </ul>
-        {this.renderTotals()}
+        <ExerciseResultsTotals/>
         {loading}
         {error}
       </section>
